build(composer): define PSR-4 autoload rule for development purposes
Classes needed to run the test suite should not be included in the main autoload
rules to avoid polluting the autoloader in production and when other people use
your package as a dependency. Therefore, it is a good idea to rely on
a dedicated path for your unit tests and to add it within the autoload-dev
section.

The CLI Ubuntu command (PHP8.2 - see the parent commit ab6e917):
- `$ /usr/bin/php /usr/bin/composer dump-autoload`

More info:
[GetComposer.org - 04 Schema - autoload-dev](https://getcomposer.org/doc/04-schema.md#autoload-dev).