build(codesniffer): add config file for coding standard
The `phpcs.xml` file has exactly the same format as a normal ruleset.xml file,
so all the same options are available in it. The `phpcs.xml` file essentially
acts as a default coding standard and configuration file for a code base, and is
typically used to allow the `phpcs` command to be run on a repository without
specifying any arguments.

More info:
[PHP_CodeSniffer - Using a Default Configuration File](https://github.com/PHPCSStandards/PHP_CodeSniffer/wiki/Advanced-Usage#using-a-default-configuration-file).